update and fix react-nlp-annotate (looks better now)
