+++ src/components/TextEntityRecognition/index.js
@@ -1,50 +1,11 @@
 import React from "react"
-import getTaskDescription from "../../utils/get-task-description.js"
-import SampleContainer from "../SampleContainer"
-import NLPAnnotator from "react-nlp-annotate/components/NLPAnnotator"
+import NLPAnnotator from "react-nlp-annotate"
 import useClobberedState from "../../utils/use-clobbered-state"
-
-const simpleSequenceToEntitySequence = (simpleSeq) => {
-  const entSeq = []
-  let charsPassed = 0
-  for (const seq of simpleSeq) {
-    if (seq.label) {
-      entSeq.push({
-        text: seq.text,
-        label: seq.label,
-        start: charsPassed,
-        end: charsPassed + seq.text.length,
-      })
-    }
-    charsPassed += seq.text.length
-  }
-  return entSeq
-}
-
-const entitySequenceToSimpleSeq = (doc, entSeq) => {
-  if (!entSeq) return undefined
-  const simpleSeq = []
-  entSeq = [...entSeq]
-  entSeq.sort((a, b) => a.start - b.start)
-  let nextEntity = 0
-  for (let i = 0; i < doc.length; i++) {
-    if ((entSeq[nextEntity] || {}).start === i) {
-      simpleSeq.push({
-        text: entSeq[nextEntity].text,
-        label: entSeq[nextEntity].label,
-      })
-      i = entSeq[nextEntity].end
-      nextEntity += 1
-    } else {
-      if (simpleSeq.length === 0 || simpleSeq[simpleSeq.length - 1].label) {
-        simpleSeq.push({ text: doc[i] })
-      } else {
-        simpleSeq[simpleSeq.length - 1].text += doc[i]
-      }
-    }
-  }
-  return simpleSeq
-}
+import Box from "@material-ui/core/Box"
+import {
+  simpleSequenceToEntitySequence,
+  entitySequenceToSimpleSeq,
+} from "./convert-react-nlp-annotate-types"
 
 export const TextEntityRecognition = (props) => {
   const [currentSampleIndex, setCurrentSampleIndex] = useClobberedState(
@@ -63,40 +24,40 @@ export const TextEntityRecognition = (props) => {
   }
 
   return (
-    <SampleContainer
-      {...props.containerProps}
-      currentSampleIndex={currentSampleIndex}
-      totalSamples={props.samples.length}
-      taskOutput={props.samples.map((s) => s.annotation)}
-      description={
-        getTaskDescription(props.samples[currentSampleIndex]) ||
-        props.interface.description
-      }
-      onChangeSample={(sampleIndex) => {
-        if (props.containerProps.onExit) {
-          props.containerProps.onExit(
-            sampleIndex > currentSampleIndex ? "go-to-next" : "go-to-previous"
-          )
+    <NLPAnnotator
+      key={currentSampleIndex}
+      titleContent={<Box paddingLeft={4}>Sample {currentSampleIndex}</Box>}
+      type="label-sequence"
+      document={props.samples[currentSampleIndex].document}
+      labels={props.interface.labels || props.interface.availableLabels}
+      initialSequence={initialSequence}
+      onPrev={(result) => {
+        props.onSaveTaskOutputItem(currentSampleIndex, {
+          entities: simpleSequenceToEntitySequence(result),
+        })
+        if (setCurrentSampleIndex) {
+          setCurrentSampleIndex(currentSampleIndex - 1)
         } else {
-          setCurrentSampleIndex(sampleIndex)
+          props.onExit("go-to-previous")
         }
       }}
-    >
-      <NLPAnnotator
-        key={currentSampleIndex}
-        type="label-sequence"
-        document={props.samples[currentSampleIndex].document}
-        labels={props.interface.labels || props.interface.availableLabels}
-        initialSequence={initialSequence}
-        onFinish={(result) => {
-          props.onSaveTaskOutputItem(currentSampleIndex, {
-            entities: simpleSequenceToEntitySequence(result),
-          })
-          if (props.containerProps.onExit)
-            props.containerProps.onExit("go-to-next")
-        }}
-      />
-    </SampleContainer>
+      onNext={(result) => {
+        props.onSaveTaskOutputItem(currentSampleIndex, {
+          entities: simpleSequenceToEntitySequence(result),
+        })
+        if (setCurrentSampleIndex) {
+          setCurrentSampleIndex(currentSampleIndex + 1)
+        } else {
+          props.onExit("go-to-next")
+        }
+      }}
+      onFinish={(result) => {
+        props.onSaveTaskOutputItem(currentSampleIndex, {
+          entities: simpleSequenceToEntitySequence(result),
+        })
+        props.onExit()
+      }}
+    />
   )
 }
